# api/generated.go
// Package api provides primitives to interact the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/deepmap/oapi-codegen/pkg/runtime"
	"github.com/labstack/echo/v4"
)

// CreateSessionRequest defines model for CreateSessionRequest.
type CreateSessionRequest struct {
	Password string `json:"password"`
	Username string `json:"username"`
}

// CreateSessionResponse defines model for CreateSessionResponse.
type CreateSessionResponse struct {
	Token string `json:"token"`
}

// Customer defines model for Customer.
type Customer struct {

	// The customer DID
	Did *string `json:"did,omitempty"`

	// The internal customer ID
	Id string `json:"id"`

	// Internal name for this customer
	Name string `json:"name"`
}

// CustomersResponse defines model for CustomersResponse.
type CustomersResponse []Customer

// ServiceProvider defines model for ServiceProvider.
type ServiceProvider struct {

	// Email addres available for other service providers in the network for getting support
	Email string `json:"email"`

	// The DID of the service provider
	Id string `json:"id"`

	// The name of the service provider
	Name string `json:"name"`

	// Number available for other service providers in the network to call in case of emergency
	Phone string `json:"phone"`
}

// CreateSessionJSONBody defines parameters for CreateSession.
type CreateSessionJSONBody CreateSessionRequest

// CreateServiceProviderJSONBody defines parameters for CreateServiceProvider.
type CreateServiceProviderJSONBody ServiceProvider

// UpdateServiceProviderJSONBody defines parameters for UpdateServiceProvider.
type UpdateServiceProviderJSONBody ServiceProvider

// CreateSessionRequestBody defines body for CreateSession for application/json ContentType.
type CreateSessionJSONRequestBody CreateSessionJSONBody

// CreateServiceProviderRequestBody defines body for CreateServiceProvider for application/json ContentType.
type CreateServiceProviderJSONRequestBody CreateServiceProviderJSONBody

// UpdateServiceProviderRequestBody defines body for UpdateServiceProvider for application/json ContentType.
type UpdateServiceProviderJSONRequestBody UpdateServiceProviderJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /web/auth)
	CreateSession(ctx echo.Context) error

	// (POST /web/customer/{id}/connect)
	ConnectCustomer(ctx echo.Context, id string) error

	// (GET /web/customers)
	GetCustomers(ctx echo.Context) error

	// (GET /web/service-provider)
	GetServiceProvider(ctx echo.Context) error

	// (POST /web/service-provider)
	CreateServiceProvider(ctx echo.Context) error

	// (PUT /web/service-provider)
	UpdateServiceProvider(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// CreateSession converts echo context to params.
func (w *ServerInterfaceWrapper) CreateSession(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.CreateSession(ctx)
	return err
}

// ConnectCustomer converts echo context to params.
func (w *ServerInterfaceWrapper) ConnectCustomer(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameter("simple", false, "id", ctx.Param("id"), &id)
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.ConnectCustomer(ctx, id)
	return err
}

// GetCustomers converts echo context to params.
func (w *ServerInterfaceWrapper) GetCustomers(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetCustomers(ctx)
	return err
}

// GetServiceProvider converts echo context to params.
func (w *ServerInterfaceWrapper) GetServiceProvider(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.GetServiceProvider(ctx)
	return err
}

// CreateServiceProvider converts echo context to params.
func (w *ServerInterfaceWrapper) CreateServiceProvider(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.CreateServiceProvider(ctx)
	return err
}

// UpdateServiceProvider converts echo context to params.
func (w *ServerInterfaceWrapper) UpdateServiceProvider(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshalled arguments
	err = w.Handler.UpdateServiceProvider(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/web/auth", wrapper.CreateSession)
	router.POST(baseURL+"/web/customer/:id/connect", wrapper.ConnectCustomer)
	router.GET(baseURL+"/web/customers", wrapper.GetCustomers)
	router.GET(baseURL+"/web/service-provider", wrapper.GetServiceProvider)
	router.POST(baseURL+"/web/service-provider", wrapper.CreateServiceProvider)
	router.PUT(baseURL+"/web/service-provider", wrapper.UpdateServiceProvider)

}